backend/requirements.txt
google-genai
fastapi
uvicorn
python-dotenv
pydantic
beautifulsoup4
httpx
